refactor: extract controllers and create barrel index
Move business logic from index.ts to controllers.ts and convert
index.ts to a barrel file that re-exports from controllers and
services. Delete old index.tsx entry point. Update dom-utils
to use correct module imports.
